requirements.txt
# Core dependencies
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3


